🚚 rename: AuthRoute 위치를 common으로 바꾸고 가져오기 업데이트 #36
